requirements.txt
alembic==1.13.1
astunparse==1.6.3
Babel==2.15.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MQTT==1.2.1
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.24.4
paho-mqtt==1.6.1
pandas==2.0.3
passlib==1.7.4
pdoc==14.6.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.18.2
